[kong]: small refactoring
