+++ krates/kong/src/lib.rs
@@ -17,6 +17,7 @@ mod error;
 mod error_response;
 pub mod inputs;
 mod konfig;
+mod kontrol;
 mod kroute;
 pub mod log;
 pub mod validate;
@@ -24,18 +25,16 @@ pub mod validate;
 pub use error::KError;
 pub use error_response::ErrorResponse;
 pub use konfig::Konfig;
-pub use kroute::kroute;
+pub use kontrol::Kontrol;
+pub use kroute::{kroute, Method};
 pub use krypto;
 pub use rouille as server;
-pub use serde_json::error::Error as JsonError;
-pub use serde_json::from_str as json_from_str;
-pub use serde_json::json;
-pub use serde_json::value::Value as JsonValue;
+pub use serde_json::{
+    error::Error as JsonError, from_str as json_from_str, json, Value as JsonValue,
+};
 
 use krypto::kpassport::Kpassport;
-use rouille::{Request, Response};
-use route_recognizer::{Params, Router};
-use std::{fs::File, str::FromStr};
+use std::fs::File;
 
 /// 🔥 Kong object
 pub struct Kong {
@@ -52,12 +51,12 @@ impl Kong {
     /// not exist and it content if it does not exist (for example the
     /// LOG file)
     fn init(config: &Konfig) {
-        Kong::create_working_directory(config);
+        Kong::create_working_directory();
         Kong::create_log_file(config);
     }
 
     /// Create working dirctory if it does not already exist
-    fn create_working_directory(config: &Konfig) {
+    fn create_working_directory() {
         // Get path to working directory
         let working_dir = Konfig::read_working_dir();
         let working_dir = std::path::Path::new(&working_dir);
@@ -104,75 +103,6 @@ impl Default for Kong {
             config,
             kpassport: None,
             input: None,
-        }
-    }
-}
-
-/// API Enpoint kontrollers
-pub trait Kontrol {
-    /// Endpoint address
-    fn address(&self) -> String;
-    /// Enpoint method
-    fn method(&self) -> Method;
-
-    /// Get user input
-    fn get_input(&self, _request: &Request) -> Option<serde_json::Value> {
-        None
-    }
-    /// Validate user input
-    fn validate(&self, input: Option<serde_json::Value>) -> Result<Option<serde_json::Value>, ()> {
-        Ok(input)
-    }
-
-    /// Handle endpoint (business logic)
-    fn kontrol(&self, kong: &Kong) -> Response;
-
-    /// url parameters extractor
-    fn url_params(
-        &self,
-        router: &Router<fn(&Kong, &Request) -> Response>,
-        url: &str,
-    ) -> Result<Params, KError> {
-        let router = router.clone();
-        let m = router.recognize(url);
-
-        match m {
-            Ok(mtch) => Ok(mtch.params().clone()),
-            Err(_) => Err(KError::UrlParsing),
-        }
-    }
-}
-
-#[derive(Clone, PartialEq)]
-/// HTTP methods
-pub enum Method {
-    /// HTTP GET method
-    Get,
-    /// HTTP POST method
-    Post,
-    /// HTTP PUT method
-    Put,
-    /// HTTP HEAD method
-    Head,
-    /// HTTP DELETE method
-    Delete,
-    /// HTTP OPTIONS method
-    Options,
-}
-
-impl Copy for Method {}
-impl FromStr for Method {
-    type Err = KError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "GET" => Ok(Method::Get),
-            "POST" => Ok(Method::Post),
-            "PUT" => Ok(Method::Put),
-            "HEAD" => Ok(Method::Head),
-            "DELETE" => Ok(Method::Delete),
-            "OPTIONS" => Ok(Method::Options),
-            _ => Err(KError::InvalidHttpMethod),
         }
     }
 }
